Refactor folder and file strutures
