+++ go.mod
@@ -5,11 +5,13 @@ module k8s.io/sample-cli-plugin
 go 1.20
 
 require (
+	github.com/spf13/cobra v1.7.0
 	github.com/spf13/pflag v1.0.5
 	k8s.io/client-go v0.0.0
 )
 
 require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.17 // indirect
 )
