+++ internal/pkg/live_print.go
@@ -1,4 +1,4 @@
-package cmd
+package pkg
 
 import (
 	"os"
